+++ server/src/__tests__/helpers/mocks.ts
@@ -7,7 +7,7 @@
 
 const { v4: uuidv4 } = require('uuid');
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
+
 type AnyRecord = Record<string, any>;
 
 /**